backend/requirements.txt
fastapi
uvicorn
python-multipart
PyMuPDF
psycopg2-binary
python-dotenv
sqlmodel
pydantic-settings
python-docx
openai